src: projects: Add separate gnl examples for xcup and xc7 families
GNL examples use interchange netlists, which target specific device
parts. This netlist is then converted to edif using RapidWright.
Finally, it is placed and routed by Vivado.

During the conversion, Rapid write may substitute the generic primitives
from the netlist for architecture-specific primitives. For example, in
the case of GNL examples, the netlist contains IBUF primitives, but the
netlist converted to edif for US+ contains INBUF or IBUFCTRL. Due to
that, each netlist should target a specific part and be a separate
example. Hence the idea of separating XC7 and XCUP examples.

To simplify the structure of the example, all of the examples for XC7
use a netlist prepared for xc7a35tcsg324-1, which is not an ideal
solution but it lovers the number of examples added to fpga-tool-perf.
Although the netlists produced by Vivado can be different for each part
of the family, the netlist targeting small parts from the same family
should be possible to implement on other, bigger parts of the family.

Internal-tag: [#38607]
Signed-off-by: Robert Winkler <rwinkler@antmicro.com>